+++ kool-core/src/commonMain/kotlin/de/fabmax/kool/scene/Lighting.kt
@@ -1,6 +1,6 @@
 package de.fabmax.kool.scene
 
-import de.fabmax.kool.math.Mat4f
+import de.fabmax.kool.math.Mat4d
 import de.fabmax.kool.math.MutableVec3f
 import de.fabmax.kool.math.Vec3f
 import de.fabmax.kool.util.Color
@@ -26,7 +26,7 @@ class Light {
     val position = MutableVec3f()
     var spotAngle = 60f
 
-    val lightMvpMat = Mat4f()
+    val lightMvpMat = Mat4d()
 
     fun setColor(color: Color, intensity: Float): Light {
         this.color.set(color)
